fix 64bit build warnings in modules/dav
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@239422 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/dav/fs/repos.c
@@ -1988,7 +1988,7 @@ static dav_error *dav_fs_patch_validate(const dav_resource *resource,
                              "submitted is invalid.");
     }
 
-    *context = (void *)(value == 'T');
+    *context = (void *)((long)(value == 'T'));
 
     return NULL;
 
@@ -2006,9 +2006,9 @@ static dav_error *dav_fs_patch_exec(const dav_resource *resource,
                                     void *context,
                                     dav_liveprop_rollback **rollback_ctx)
 {
-    int value = context != NULL;
+    long value = context != NULL;
     apr_fileperms_t perms = resource->info->finfo.protection;
-    int old_value = (perms & APR_UEXECUTE) != 0;
+    long old_value = (perms & APR_UEXECUTE) != 0;
 
     /* assert: prop == executable. operation == SET. */
